Fix payment api endpoint url bug
There's a bug in the payment api that causes orders log as waiting for payment even though payment is successfully done. The bug is caused by conflicting rewrite rules that make the permalink wrong. I make a filter for checkout payment url and remove these extra parts in the url which makes the payment api to work.

My solution isn't pretty and could use some regex magic but it does what it needs to do and it works.

You can debug the url by adding this in header of your theme:
global $woocommerce;
$order = new WC_Order($order_id);
var_dump($order->get_checkout_payment_url());