Fix #415 order email title & subject translations
Fixes introduced by 89ca15df970ee410513200954f0993fcf0ab93f9 added a
clever comparison of the current locale with the order one to determine
whether the content should be translated or not.

get_locale helper is contextual and might not return the right locale.
This commit uses the global WP locale of the site instead.